requirements.txt
fastapi==0.115.5
uvicorn==0.32.1
python-dotenv==1.0.1
anthropic==0.80.0
pydantic>=2.0.0
aiofiles==24.1.0
httpx==0.28.1
python-multipart==0.0.12

# Security packages
slowapi==0.1.9          # Rate limiting
itsdangerous==2.2.0     # CSRF tokens
bleach==6.2.0           # XSS prevention
cryptography>=40.0.0    # Secret encryption
validators==0.34.0      # Input validation

# Testing packages
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.27.0
